1. Demonstrates the use of @OneToMany establish unidirectional one to many entity relation.
Owning party of this relationship is Transaction
Non-Owning party of this relationship is Account

But since this is a  unidirectional , Transactions doe not have reference to the Account entity.

Hence , we put the @JoinColumn on the Account entity.

Furthermore, note that the Transaction entity doe not have ACCOUNT_ID filed insted we configure @JoinColumn( ,  nullable = false)
so hibernate automatically populate the accountId filed in the Transaction.